requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
groq
python-dotenv==1.0.1
pypdf==6.4.0
langchain==0.3.27
langchain-core==0.3.77
langchain-community==0.3.30
langchain-text-splitters==0.3.11
chromadb==1.0.20
sentence-transformers==2.6.0
